% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{add_root_to_newick}
\alias{add_root_to_newick}
\title{Add a root node to a newick tree string with distance 0 from the root}
\usage{
add_root_to_newick(tree, root_name = "root")
}
\arguments{
\item{tree}{A newick tree string to add a root node to.}

\item{root_name}{The name of the root node to add. Default is "root".}
}
\description{
A utility function for placing an explicit "root" into a newick tree string.
at the implicit root. This function performs: `(tree); -> (tree:0, root:0);`.
}
\seealso{
phylogeny_to_newick
}
