% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_rates.R
\name{biphasic_HIV_rate}
\alias{biphasic_HIV_rate}
\title{Simple biphasic rate function used in [Kupperman et al.]}
\usage{
biphasic_HIV_rate(current_step, birth_step, params)
}
\arguments{
\item{current_step}{Current time step}

\item{birth_step}{When the infection was generated}

\item{params}{A list with one element `R0` needed to evaluate the rate function.}
}
\description{
Simple biphasic rate function used in [Kupperman et al.]
}
