% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_distance_matrix.R
\name{build_distance_matrix_from_df}
\alias{build_distance_matrix_from_df}
\title{Build a distance matrix from a dataframe of sequences using ape}
\usage{
build_distance_matrix_from_df(df, model = "TN93", keep_root = FALSE)
}
\description{
Use ape to compute a distance matrix using the specified distance model.
Default is the "TN93" model.
}
