% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_models.R
\name{check_rate_model}
\alias{check_rate_model}
\title{Check that a rate model is mathematically valid}
\usage{
check_rate_model(rate_model)
}
\arguments{
\item{rate_model}{A list of rate model parameters.}
}
\description{
A helper function to check that a rate model paramterization is
mathematically valid.
}
\details{
Check that all rates are positive, that the fraction of bases add up to 1,
the fraction of invariant sites is in [0, 1), and the discrete gamma model
is defined for `alpha` and `ncat`.
}
\seealso{
generate_rate_model generate_sequences
}
