% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history.R
\name{clean_up}
\alias{clean_up}
\title{Clean up the simulation state and return the relevant data.}
\usage{
clean_up(state)
}
\description{
Clean up the simulation state and return the relevant data.
}
