% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing.R
\name{contact_traced_uniform_ids}
\alias{contact_traced_uniform_ids}
\title{Obtain a sample using a iterative contact tracing with a uniform discovery rate}
\usage{
contact_traced_uniform_ids(
  active,
  parents,
  minimum_sample_size,
  p,
  max_attempts = 1
)
}
\arguments{
\item{active}{A vector of active individuals}

\item{parents}{A matrix encoding the transmission history}

\item{minimum_sample_size}{The minimum number of individuals to form a sample}

\item{p}{The probability of discovering each contact during tracing}

\item{max_attempts}{Maximum number of attempts to perform to obtain a sample.
If no sample is found after this number of attempts, `FALSE` is returned}
}
\value{
A list with three fields: "status", "samples", and "found"
}
\description{
Perform iterative contact tracing on a simulated contact network.
Randomness within contact tracing comes from the probability of discovering
a new infection. This function assumes the infection rate is uniform across
all possible contacts and individuals.
}
\details{
Contact tracing is terminated when 2 * min_sample_size active nodes are discovered,
or when there are no more detections to trace. At least `min_sample_size`
individuals will always be returned.

To determine the initial detection, we loop over a the list of active nodes
until we complete a successful contact tracing.

All contact tracing algorithms store both the ids of all discovered individuals
and the ids of discovered active individuals.
}
