% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing.R
\name{contact_tracing_core}
\alias{contact_tracing_core}
\title{Core contact tracing algorithm}
\usage{
contact_tracing_core(
  detected_id,
  active,
  parents,
  discovery_function,
  termination_function
)
}
\arguments{
\item{detected_id}{The id of the node used to start the tracing}

\item{active}{A vector of active individuals}

\item{parents}{A matrix of transmission history}

\item{discovery_function}{A function that takes a list of nodes and relative
transmission times and determines which will be included}
}
\description{
Given a set of active nodes and the transmision history (parents),
sample a set of nodes that are contact traced using the `discovery_function`
parameter to determine the transmission rate.
}
