% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing.R
\name{contact_tracing_engine}
\alias{contact_tracing_engine}
\title{Perform contact tracing on a simulated contact network}
\usage{
contact_tracing_engine(
  detected_id,
  active,
  parents,
  discovery_function,
  max_attempts,
  termination_function,
  minimum_size = 3
)
}
\arguments{
\item{detected_id}{A vector of individuals to begin tracing with}

\item{active}{A vector of active individuals}

\item{parents}{A matrix encoding the transmission history}

\item{discovery_function}{A function to determine which individuals to trace}

\item{termination_function}{A function to determine when to terminate tracing a contact tracing.}

\item{minimum_size}{The minimum number of individuals to form a sample}
}
\value{
A list with three fields: "status", "samples", and "found"
}
\description{
The engine for simulating contact tracing. Provide network information
(`active`, `parents`), a list or vector of individuals to begin tracing with
(`detected_id`), functions to control the discovery and termination
of tracing (`discovery_function`, `termination_function`), and parameters about
iteration (`max_attempts`) and fallback termination conditions
(`minimum_size`).
}
