% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{fasta_string_to_dataframe}
\alias{fasta_string_to_dataframe}
\title{Convert a fasta string output by seq-gen to a dataframe}
\usage{
fasta_string_to_dataframe(fasta_string, trim = TRUE, include_root = FALSE)
}
\value{
A dataframe with the "seq" and "name" columns
}
\description{
A utility function to build a dataframe from a fasta string.
}
\details{
@param fasta_string A string containing the fasta output from seq-gen
}
