% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history.R
\name{gen_const_phase}
\alias{gen_const_phase}
\title{Simulate a fixed number of steps}
\usage{
gen_const_phase(state, parameters, num_steps, verbose = FALSE)
}
\description{
Usually, we want to simulate a fixed number of steps after the exponential growth phase.
This function performs that step.
}
