% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history_balanced.R
\name{gen_transmission_history_balanced_tree}
\alias{gen_transmission_history_balanced_tree}
\title{Generate a transmission history for a given number of individuals under a optimally balanced tree}
\usage{
gen_transmission_history_balanced_tree(population_size, spike_root = FALSE)
}
\arguments{
\item{n}{number of individuals to have in the final layer}
}
\description{
Simulate a balanced transmission tree with a given number of individuals. If the `population_size` is a power of 2,
then the tree will be unique up to relabeling the tips. If the `population_size` is not a power of 2, then the tree
that minimizes Sackin's index is not unique. To constrain the solution, we build the tree programmatically. We
convert each leaf node to a cherry in the initial layer, to create a new layer of leaves. We repeat until we have
a `population_size` number of leaves. As a result, we will always find the shortest tree with the prescribed number
of leaves that is perfectly balanced.
}
