% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history.R
\name{gen_transmission_history_exponential_constant}
\alias{gen_transmission_history_exponential_constant}
\title{Generate transmission history for full population}
\usage{
gen_transmission_history_exponential_constant(
  minimum_population,
  offspring_rate_fn,
  maximum_population_target,
  total_steps,
  spike_root = FALSE
)
}
\description{
Forward stochastic simulation to generate transmission history for a
complete population with an exponential growth then constant population structure.
}
\details{
This simulation is intended for HIV, but may be broadly applicable if the lifespan distribution
is adjusted to other distributions/parameters.
}
