% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_distance_matrix.R
\name{geneology_to_distance_matrix}
\alias{geneology_to_distance_matrix}
\title{Reduce a geneology to a pairwise distance matrix.}
\usage{
geneology_to_distance_matrix(geneology, mode = "mu", spike_root = FALSE)
}
\description{
Reduce a geneology to a pairwise distance matrix.
}
