% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_distance_matrix.R
\name{geneology_to_distance_matrix_classic}
\alias{geneology_to_distance_matrix_classic}
\title{Reduce a geneology to a pairwise distance matrix.}
\usage{
geneology_to_distance_matrix_classic(geneology, spike_root = FALSE)
}
\description{
Provided for backwards compatability with [Kupperman et al 2022]. New code
should use `[geneology_to_distance_matrix]`.
}
\seealso{
geneology_to_distance_matrix
}
