% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioPhyBreakInterface.R
\name{geneology_to_phylogeny_bpb}
\alias{geneology_to_phylogeny_bpb}
\title{Generate a phylogeny from a transmission history using BioPhyBreak's coalescent simulator}
\usage{
geneology_to_phylogeny_bpb(
  transmission_history,
  infection_times,
  leaf_sample_ids,
  sample_times,
  a = 5,
  b = 5,
  make_plot = FALSE
)
}
\arguments{
\item{transmission_history}{A transmission history matrix}

\item{infection_times}{A vector of infection times}
}
\description{
Run a backwards simulation using the transmission history to generate a phylogeny.
See `geneology_to_phylogeny_bpb` for details on how the input should be structured.
This function assumes that all coalescent events occur at or after the initial infection,
that there is a single introduction and ancestral sequence.
}
\seealso{
reduce_transmission_history_bpb

generate_sequences
}
