% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_models.R
\name{generate_rate_model}
\alias{generate_rate_model}
\title{Generate a rate model from provided paramters}
\usage{
generate_rate_model(
  a2c,
  a2g,
  a2t,
  c2g,
  c2t,
  g2t,
  fa,
  fc,
  fg,
  ft,
  i,
  alpha,
  ncat
)
}
\arguments{
\item{a2c}{The rate of adenosine to cytosine transversions.}

\item{a2g}{The rate of adenosine to cytosine transitions.}

\item{a2t}{The rate of adenosine to thymine transversions.}

\item{c2g}{The rate of cytosine to guanine transversions.}

\item{c2t}{The rate of cytosine to thymine transitions.}

\item{g2t}{The rate of guanine to thymine transversions.}

\item{fa}{The fraction of adenosine at equilibrum.}

\item{fc}{The fraction of cytosine at equilibrum.}

\item{fg}{The fraction of guanine at equilibrum.}

\item{ft}{The fraction of thymine at equilibrum.}

\item{i}{The proportion of invariant sites. Traditionally estimated by the
proportion of sites with no observed mutations.}

\item{alpha}{The shape parameter for the discrete gamma distribution.}

\item{ncat}{The number of categories in the discrete gamma distribution.}
}
\value{
A list of rate model parameters.
}
\description{
Accept a set of parameters for GTR+I+G model. See below for details of the
parameterization. To disable I, set the parameter i to 0. To disable G, set
alpha=1 and ncat=1 for an exponential distribution.
}
\details{
For an overview of GTR+I+G and substitition models, see
[here](https://www.ccg.unam.mx/~vinuesa/Model_fitting_in_phylogenetics.html).
For a more detailed construction, see
[Yang 1994](https://doi.org/10.1007/BF00178256), and
[Yang 1996](https://doi.org/10.1093/oxfordjournals.molbev.a025625).
}
\examples{
rate_model <- generate_rate_model(
    a2c = 1, a2g = 1, a2t = 1,
    c2g = 1, c2t = 1, g2t = 2,
    fa = 0.25, fc = 0.25,
    fg = 0.25, ft = 0.25,
    i = 0.1, alpha = 0.25, ncat = 8
)

}
\seealso{
generate_sequences get_V3_rate_model get_p17_rate_model check_rate_model
}
