% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sequences.R
\name{generate_sequences}
\alias{generate_sequences}
\title{Generate sequences from a phylogeny using Seq-Gen through phyclust}
\usage{
generate_sequences(
  phylogeny,
  branch_rate,
  root_sequence,
  rng_seed = -1,
  rate_model,
  rate_per_nt = FALSE
)
}
\arguments{
\item{phylogeny}{A phylogeny object}

\item{root_sequence}{A root sequence}

\item{rate_model}{A list of GTR+I+G model parameters. Expects a list
 of 13 parameters:  (6 rate parameters) `a2c`, `a2g`, `a2t`, `c2g`,
`c2t`, `g2t`, (nucleotide frequencies:) `fa, fc, fg, ft`,
(proportion of sites with no variation:) `i`, (site specific
heterogeneity shape parameter) `alpha`, and number of categories for
discretized gamma heterogeneity (`ncat`).}

\item{rate_per_nt}{Flag to indicate whether the mutation rate is
  per nucleotide (nt) or per sequence. If per-sequence (`rate_per_nt=FALSE`),
the mutation rate is normalized by the length of the root sequence.}
}
\description{
Given a phylogeny, generate possible sequences using Seq-Gen [Rambaut & Grassley, 1997].
A root sequence for the simulation is required. The root sequence is placed
at the MRCA of the phylogeny unless `spike_root = TRUE` is specified when the phylogeny
is constructed.
}
\details{
If the root sequence is a character vector, it is flattened into a single string.
}
\seealso{
geneology_to_phylogeny_bpb
}
