% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_rates.R
\name{get_biphasic_HIV_rate}
\alias{get_biphasic_HIV_rate}
\title{Biphasic rate function factory}
\usage{
get_biphasic_HIV_rate(params)
}
\arguments{
\item{params}{list of parameters used.}
}
\description{
Biphasic rate function factory
}
