% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_rates.R
\name{get_biphasic_HIV_rate_function}
\alias{get_biphasic_HIV_rate_function}
\title{Factory function to generate biphasic rate functions}
\usage{
get_biphasic_HIV_rate_function(
  front_density_factor,
  front_cutoff,
  target_length
)
}
\arguments{
\item{front_density_factor}{How much relative significance to place in the
first phase of the rate function.}

\item{front_cutoff}{Length of first phase. An integer 1 or greater.}

\item{target_length}{Expected length of an infection. Used for normalization
to ensure an average of $R0$ secondary infections.}
}
\value{
Callable, rate function
}
\description{
Factory function to generate biphasic rate functions
}
