% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transmission_rates.R
\name{get_Kphasic_hiv_rate_function}
\alias{get_Kphasic_hiv_rate_function}
\title{Kphasic HIV rate function factory
Provide a list of relative importance (multiples over a "base" rate) for}
\usage{
get_Kphasic_hiv_rate_function(rate_list, target_length, params)
}
\arguments{
\item{rate_list}{# List of relative rates for each phase.}

\item{target_length}{List of lengths of each phase.}

\item{params}{list with element R0}
}
\value{
Callable, rate function
}
\description{
Kphasic HIV rate function factory
Provide a list of relative importance (multiples over a "base" rate) for
}
