% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_models.R
\name{get_p17_rate_model}
\alias{get_p17_rate_model}
\title{Rate model for gag p17 estimated from the Swedish transmission chain}
\usage{
get_p17_rate_model()
}
\value{
A list of rate model parameters.
}
\description{
Estimates for GTR+I+G rate model parameter from [Leitner et al. 1997] on the
Swedish transmission chain on p17. Rate estimates for p17 follow from a
secondary analysis.
}
\examples{
rate_model <- get_p17_rate_model()
}
\seealso{
generate_sequences generate_rate_model
}
