% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_models.R
\name{get_pol_rate_model}
\alias{get_pol_rate_model}
\title{Rate model for pol estimated from the Swedish transmission chain}
\usage{
get_pol_rate_model(nonzero_I = FALSE)
}
\arguments{
\item{nonzero_I}{Boolean indicating whether to use the I parameter (invariant
sites) or not. Default is `TRUE`(0.255). If `FALSE`, the I parameter is set to 0.}
}
\value{
A list of rate model parameters.
}
\description{
Estimates for GTR+I+G rate model parameter on the
Swedish transmission chain on pol region. Follow up analysis by Lundgren et al.
[2022] on the Swedish transmission chain on pol region.
}
\seealso{
generate_sequences generate_rate_model
}
