% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rate_models.R
\name{get_V3_rate_model}
\alias{get_V3_rate_model}
\title{Rate model for env V3 estimated from the Swedish transmission chain}
\usage{
get_V3_rate_model(nonzero_I = TRUE)
}
\arguments{
\item{nonzero_I}{Boolean indicating whether to use the I parameter (invariant
sites) or not. Default is `TRUE`(0.68). If `FALSE`, the I parameter is set to 0.}
}
\value{
A list of rate model parameters.
}
\description{
Estimates for GTR+I+G rate model parameter from [Leitner et al. 1997] on the
Swedish transmission chain on envelope V3 region.
}
\examples{
rate_model <- get_V3_rate_model()

}
\seealso{
generate_sequences generate_rate_model
}
