% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history.R
\name{keep_samples}
\alias{keep_samples}
\title{Keep a set of samples in the simulation}
\usage{
keep_samples(state, samples)
}
\description{
Remove all other samples from the simulation. Intended to be used to represent a masking event,
Where only a subset of the population is propagated forward in time.
}
\details{
Provide a state object and a vector of samples to keep. A new state object is returned with only the samples kept.
}
