% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_sequences.R
\name{lookup_sequence_by_index}
\alias{lookup_sequence_by_index}
\title{Obtain a sequence from the builtin reference sequences using coordinates}
\usage{
lookup_sequence_by_index(organism_name, start, stop)
}
\arguments{
\item{start}{The start coordinate of the sequence}

\item{end}{The end coordinate of the sequence}
}
\description{
Provide an interval and a reference sequence name.
Currently supported are: "HIV1" (HXB2 reference) and "toy" (a poly A/C/G/T
example sequence for testing).
}
