% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_sequences.R
\name{lookup_sequence_by_name}
\alias{lookup_sequence_by_name}
\title{Obtain a sequence from the builtin reference sequences using an name and annotated region}
\usage{
lookup_sequence_by_name(organism_name, region_name)
}
\description{
Return a portion of a reference genome using a standard annotation name.
}
\details{
Currently supported are: "HIV1" (HXB2 reference) and "toy" (a poly A/C/G/T
example sequence for testing). Supported HIV1 regions include the short annotated
list, see (here)[https://www.hiv.lanl.gov/components/sequence/HIV/search/help.html#region]
for the full list. Clinical regions (p17, V3) are supported as "p17-clinical" and
"v3-clinical".

For a more detailed lookup proceedure with the reference sequences using user-provided
coordinates, see `lookup_sequence_by_index`.
}
\seealso{
lookup_sequence_by_index
}
