% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing.R
\name{never_terminate_early_factory}
\alias{never_terminate_early_factory}
\title{Factory function to never terminate contact tracing until all known nodes
  have been traced}
\usage{
never_terminate_early_factory()
}
\description{
When contact tracing, we don't want to stop until we have traced all known.
This is computationally more expensive, but better describes reality.
}
