% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_utils.R
\name{phylogeny_to_newick}
\alias{phylogeny_to_newick}
\title{Convert a phylogeny array to a newick tree}
\usage{
phylogeny_to_newick(phylogeny, mode = "mu", label_mode = "local")
}
\arguments{
\item{phylogeny}{A phylogeny or geneology array.}

\item{mode}{String to determine reconstruction mode. Default is "mu", for
mutation count. Alternative "mean" for expected distance.}

\item{label_mode}{String to determine how to label nodes. Default is "local",
for node index. Alternative "abs" for absolute index.}
}
\description{
Build a newick tree from a phylogeny or geneology array.
 Nodes are named "#_". Specify the mode argument to select the branch
length mode. Mode "mu" denotes a (sampled) mutation count, while mode
"mean" denotes expected distances.
}
