% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_ids.R
\name{random_fixed_size_ids}
\alias{random_fixed_size_ids}
\title{Sample a fixed number of of individuals randomly from the population.}
\usage{
random_fixed_size_ids(active, minimum_size, spike_root = FALSE)
}
\arguments{
\item{active}{A vector of active individuals}

\item{minimum_size}{The minimum number of individuals to sample.}

\item{spike_root}{Whether to include the root in the sample.}
}
\description{
A helper function for sampling a fixed number of individuals from the
population. Select a fixed number of individuals from the active set.
Does not use any information about the transmission history or the
active population size to select the sample.
}
\seealso{
random_prop_ids
}
