% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_ids.R
\name{random_prop_ids}
\alias{random_prop_ids}
\title{Sample a proportion of the active individuals}
\usage{
random_prop_ids(active, proportion, minimum_size, spike_root = FALSE)
}
\arguments{
\item{active}{A vector of active individuals}

\item{proportion}{A float between 0 and 1.
The proportion of active individuals to sample.}

\item{minimum_size}{The minimum number of individuals to sample.
If the proportional sample is smaller than the minimum size,
the proprotional size is used.}

\item{spike_root}{Whether to include the root in the sample.}
}
\description{
A helper function for sampling a proportion of the active individuals.
Select a proportion of individuals from the active set. Does not use
any information about the transmission history to select the sample.
}
\seealso{
random_fixed_size_ids
}
