% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_large_matrix.R
\name{reduce_large_matrix}
\alias{reduce_large_matrix}
\title{Reduce a large matrix by randomly sampling a cluster of closely related individuals}
\usage{
reduce_large_matrix(
  oversampled_matrix,
  subsample_size,
  spike_root = FALSE,
  root_position = 0,
  index_id = -1,
  sort_order = NULL
)
}
\description{
An obtained sample (through contact tracing or random sampling) may be larger than needed.
This function extracts a subset of closely related individuals with a randomly selected "center".
This respects the `spike_root` option, and if `spike_root = TRUE`, will return the root individual
in the last column.
}
