% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_transmission_history_to_true_geneology.R
\name{reduce_transmission_history}
\alias{reduce_transmission_history}
\title{Reduce simulation output to transmission history for a subset.}
\usage{
reduce_transmission_history(samples, parents, current_step, spike_root = FALSE)
}
\arguments{
\item{samples}{A vector of individuals (integers) to include in the sample.}

\item{parents}{A matrix of parental individuals that encodes the transmission
history and sample times.}

\item{current_step}{The current (absolute) time step in the simulation.}

\item{spike_root}{A boolean indicating whether the geneology should
 include the root of the outbreak or not. Default is `FALSE`.
 This should be specified even if the founding infection is sampled,
as the root of the outbreak will have evolved since the founding event.}
}
\value{
A list with 1 element: "geneology" a matrix of transmission history
  that encodes an evolutionary tree.
}
\description{
Reduce a large simulation output to a smaller transmission history for a subset
by tracing back the ancestory of each individual in the sample. If `spike_root`
is `TRUE`, then the root of the tree is included in the geneology.
We call this a geneology, rather than a phylogeny as it assumes that the
transmission history is the phylogeny and no within-host diversity occurs.
}
\details{
To include within host diversity, use `reduce_transmission_history_bpb` to
extract a subst of the transmission history that we need for the phylogeny,
and see  `geneology_to_phylogeny_bpb` to simulate within-host diversity
and recover a true phylogeny.
}
\seealso{
reduce_transmission_history_bpb
}
