% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history.R
\name{remove_samples}
\alias{remove_samples}
\title{Remove samples from the simulation}
\usage{
remove_samples(state, samples)
}
\description{
Remove a set of samples from the simulation. Intended to be used with contact tracing, we expect
individuals identified through contact tracing to be removed from the population of active (uncontrolled or undiagnosed)
cases.
}
\details{
Provide a state object and a vector of samples to remove. A new state object is returned with the samples removed.
}
