% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_sequences.R
\name{show_available_regions}
\alias{show_available_regions}
\title{Show the list of available regions for a provided reference sequence}
\usage{
show_available_regions(organism_name, chart = TRUE)
}
\arguments{
\item{organism_name}{The name of the reference sequence}

\item{chart}{bool (default TRUE). If TRUE, a chart of the available regions
is printed to the screen.}
}
\value{
A list of available regions for the named reference sequence.
If the organism_name is not found, an NULL is printed.
}
\description{
Show the list of available regions for a provided reference sequence
}
