% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_sequences.R
\name{show_available_sequences}
\alias{show_available_sequences}
\title{Show the list of available reference sequences}
\usage{
show_available_sequences()
}
\description{
A list of all available organism/sequence names is printed to the screen.
}
