% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_triple_HIV_V3.R
\name{simulate_all_paradigms_HIV_V3}
\alias{simulate_all_paradigms_HIV_V3}
\title{Sample distance matrices for all three major simulation paradigms:
the transmission history, the phylogeny, and derived from simulated sequences (TN93)}
\usage{
simulate_all_paradigms_HIV_V3(params)
}
\value{
list with keys "matrix" and "input_params"
}
\description{
This simulation returns three distance matrices for the same set of individuals.
The first matrix (matrix_trans) is derived from the transmission history.
The second matrix (matrix_phylo) is derived from the phylogeny by simulating a
coalescent process on the transmission history.
The third matrix (matrix_seq) is derived from the sequences by computing the
pairwise distances using the TN93 model.
}
\examples{
parameters <- list("rate_function_parameters" = list("R0"=5), "a"=5, "b"=5,
                    "mutation_rate" = 0.0067,  # units are per nt
                    "contact_tracing_discovery_probability" = 0.9,
                    "minimum_population"=15, "maximum_population_target"=500)

}
