% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_classic_HIV.R
\name{simulate_classic_HIV}
\alias{simulate_classic_HIV}
\title{Classic HIV simulator for population level simulations.}
\usage{
simulate_classic_HIV(params)
}
\value{
list with key "matrix"
}
\description{
A simple agent-based simulation for HIV. Returns pairwise distance matrix samples
obtained from a simulated outbreak. This simulation models population level diversity,
but does not explicitly consider within-host diversity. A mutation clock rate, R0, and
several epidemiological parameters are required to describe the duration of infections.
}
\examples{
parameters <- list("rate_function_parameters" <- list("R0"=5),
                    "mutation_rate" = 0.0067 / 12 * 300,
                    "minimum_population"=50, "maximum_population_target"=100)

}
