% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_modern_HIV.R
\name{simulate_modern_HIV}
\alias{simulate_modern_HIV}
\title{Modern HIV simulator}
\usage{
simulate_modern_HIV(params)
}
\value{
list with keys "matrix" and "input_params"
}
\description{
A modern agent-based simulation for HIV that explicitly models within-host simulations using biophybreak extensions.
}
\examples{
parameters <- list("rate_function_parameters" = list("R0"=5), "a"=5, "b"=5,
                    "minimum_population"=15, "maximum_population_target"=500)

}
