% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_sequences_HIV_V3.R
\name{simulate_sequences_HIV_V3}
\alias{simulate_sequences_HIV_V3}
\title{Modern HIV epidemic sequences simulator}
\usage{
simulate_sequences_HIV_V3(params)
}
\value{
list with keys "matrix" and "input_params"
}
\description{
A modern agent-based simulation for HIV that explicitly models within-host simulations using biophybreak extensions.
This pre-build simulation returns a collection of sequences, rather than a matrix.
}
\examples{
parameters <- list("rate_function_parameters" = list("R0"=5), "a"=5, "b"=5,
                    "mutation_rate" = 0.0067,  # units are per nt
                    "contact_tracing_discovery_probability" = 0.9,
                    "minimum_population"=15, "maximum_population_target"=500)

}
