% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastify_transmission_history.R
\name{stochastify_transmission_history}
\alias{stochastify_transmission_history}
\title{Sample evolutionary distances to an integer number of mutations per branch}
\usage{
stochastify_transmission_history(transmission_history, rate)
}
\arguments{
\item{transmission_history}{A matrix output by `reduce transmission history`.}

\item{rate}{Clock rate for evolutionary process with units of substitions per sequence per year.}
}
\description{
Convert temporal distances for a geneology to an integer number of mutations.
This step is extracted so a single geneology can be re-sampled and create multiple possible matrices.
}
