% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_tracing.R
\name{sufficient_data_data_factory}
\alias{sufficient_data_data_factory}
\title{Factory function for a detection to check termination condition based on
A sufficient amount of individuals
Terminate when we have found a `minimum_size` number of active individuals}
\usage{
sufficient_data_data_factory(minimum_size)
}
\arguments{
\item{minimum_size}{The minimum number of discovered active individuals to
terminate}
}
\description{
Factory function for a detection to check termination condition based on
A sufficient amount of individuals
Terminate when we have found a `minimum_size` number of active individuals
}
