% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_transmission_history.R
\name{validate_state}
\alias{validate_state}
\title{Utility function to check that a state is (computationally) valid}
\usage{
validate_state(state)
}
\description{
Check that the state object is a valid and has the correct list elements.
}
